+++ android-wms/app/src/main/java/com/akan/wms/mvp/fragment/manager/OutTransferAddNewFragment.java
@@ -424,7 +424,7 @@ public class OutTransferAddNewFragment extends BaseFragment<IOutTransferView, Ou
                 scanBean.setItem_id(detail.getItem_id());
                 scanBean.setItem_spec(detail.getItem_spec());
                 scanBean.setItem_name(detail.getItem_name());
-                scanBean.setSend_qty(detail.getApply_qty());
+                scanBean.setSend_qty((int) detail.getApply_qty());
                 scanBean.setArrive_qty(detail.getPoint_qty());
                 scanBean.setBarList(item.getBarList());//历史条码
                 scanList.add(scanBean);
